pyproject.toml
[project]
name = "aporia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "pydantic>=2.4.0",
    "pydantic-ai>=0.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "streamlit>=1.27.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=8.0.0",
]
